src/user-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
sqlalchemy
alembic
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
requests

# Testing
pytest
